cogl/onscreen/glx: Fix x/y mixup used to identify the current output
I haven't seen this cause a problem but it looks like the worst case is
that it would have put the wrong refresh rate value in CoglFrameInfo
for multi-monitor systems of differing frequencies. But even that seems
unlikely given `_cogl_xlib_renderer_output_for_rectangle` chooses the
output with the greatest overlap of the partially-correct rectangle.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/3569>